+++ src/main/java/qouteall/imm_ptl/core/mixin/common/collision/MixinAbstractArrow.java
@@ -1,6 +1,6 @@
 package qouteall.imm_ptl.core.mixin.common.collision;
 
-import net.minecraft.world.entity.projectile.AbstractArrow;
+import net.minecraft.world.entity.projectile.arrow.AbstractArrow;
 import org.spongepowered.asm.mixin.Mixin;
 
 @Mixin(AbstractArrow.class)