+++ src/main/java/qouteall/imm_ptl/core/mixin/common/collision/MixinThrownEnderPearl.java
@@ -3,7 +3,7 @@ package qouteall.imm_ptl.core.mixin.common.collision;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.Entity;
-import net.minecraft.world.entity.projectile.ThrownEnderpearl;
+import net.minecraft.world.entity.projectile.throwableitemprojectile.ThrownEnderpearl;
 import net.minecraft.world.phys.HitResult;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;