+++ src/main/java/qouteall/q_misc_util/mixin/MixinMinecraftServer_Misc.java
@@ -4,8 +4,7 @@ import com.mojang.datafixers.DataFixer;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.Services;
 import net.minecraft.server.WorldStem;
-import net.minecraft.server.level.progress.ChunkProgressListener;
-import net.minecraft.server.level.progress.ChunkProgressListenerFactory;
+import net.minecraft.server.level.progress.LevelLoadListener;
 import net.minecraft.server.packs.repository.PackRepository;
 import net.minecraft.util.thread.ReentrantBlockableEventLoop;
 import net.minecraft.world.level.storage.LevelStorageSource;
@@ -50,13 +49,13 @@ public abstract class MixinMinecraftServer_Misc extends ReentrantBlockableEventL
         at = @At("RETURN")
     )
     private void onConstruct(
-        Thread thread, LevelStorageSource.LevelStorageAccess levelStorageAccess, PackRepository packRepository, WorldStem worldStem, Proxy proxy, DataFixer dataFixer, Services services, ChunkProgressListenerFactory chunkProgressListenerFactory, CallbackInfo ci
+        Thread thread, LevelStorageSource.LevelStorageAccess levelStorageAccess, PackRepository packRepository, WorldStem worldStem, Proxy proxy, DataFixer dataFixer, Services services, LevelLoadListener levelLoadListener, CallbackInfo ci
     ) {
         MiscGlobals.refMinecraftServer = new WeakReference<>((MinecraftServer) ((Object) this));
     }
     
     @Inject(method = "createLevels", at = @At("RETURN"))
-    private void onWorldsCreated(ChunkProgressListener listener, CallbackInfo ci) {
+    private void onWorldsCreated(CallbackInfo ci) {
         DimensionIntId.onServerStarted((MinecraftServer) (Object) this);
     }